GH-10: Moved models into separate model folder.
